// src/app/shared/models/enums/alias-type.enum.ts
export enum AliasType {
  ROOM_NAME = 'ROOM_NAME',
  H323_E164 = 'H323_E164',
  /**
   * Represents H323 URI.
   */
  H323_URI = 'H323_URI',
  /**
   * Represents H323 IP.
   */
  H323_IP = 'H323_IP',
  SIP_URI = 'SIP_URI',
  /**
   * Represents SIP IP.
   */
  SIP_IP = 'SIP_IP',
  /**
   * Represents guest dial string for SW client LifeSize ClearSea
   */
  CS_DIAL_STRING = 'CS_DIAL_STRING',
  /**
   * Represents room URL for ADOBE_CONNECT.
   */
  ADOBE_CONNECT_URI = 'ADOBE_CONNECT_URI',
  /**
   * Represents conference number for FREEPBX.
   */
  FREEPBX_CONFERENCE_NUMBER = 'FREEPBX_CONFERENCE_NUMBER',
  SKYPE_URI = 'SKYPE_URI',
  RTMP_NAME = 'RTMP_NAME',
  WEB_RTC_NAME = 'WEB_RTC_NAME',
  /**
   * Represents URL for the Pexip web client.
   */
  WEB_CLIENT_URI = 'WEB_CLIENT_URI',
  ROOM_NUMBER = 'ROOM_NUMBER',
  PEXIP_ROOM_NUMBER_URI = 'PEXIP_ROOM_NUMBER_URI',
  PEXIP_PHONE_NUMBER_URI = 'PEXIP_PHONE_NUMBER_URI',
}